shops/forms/base.py
from django import forms
from django.core.exceptions import ValidationError
from datetime import datetime
from crispy_forms.helper import FormHelper
from core_operations.constants import LIST_OF_STATES_IN_US
from crispy_forms.layout import Layout, Fieldset, Submit, Field, ButtonHolder, HTML, Reset, Column, Row, Div, Button, Hidden
from crispy_forms.bootstrap import FormActions, InlineCheckboxes, InlineField
from django.urls import reverse
from django_recaptcha.fields import ReCaptchaField
from  django_recaptcha.widgets import ReCaptchaV2Invisible, ReCaptchaV2Checkbox, ReCaptchaV3