requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
# attrs==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
billiard==3.6.4.0
blinker==1.6.2
beautifulsoup4==4.12.2
CacheControl==0.12.11
cachetools==5.3.0
captcha==0.4
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
channels-redis==4.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
crispy-forms-foundation==1.0.2
cron-descriptor==1.4.0
cryptography==41.0.2
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
dj-stripe==2.8.3

Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
django-versatileimagefield==3.0
django-formtools==2.4.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
# easyocr==1.7.1  #torch dependency. computer vision related
et-xmlfile==1.1.0
Faker==19.6.2
filelock==3.12.2
# filterpy==1.4.5
firebase-admin==6.2.0
fredapi==0.5.1
frozenlist==1.4.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-recaptcha==2.0.1
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
googlemaps==4.10.0
grpcio==1.53.0
grpcio-status==1.53.0
holidays>=0.33
# HeapDict==1.0.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
# iPython==8.14.0
itsdangerous==2.1.2
# Jinja2==3.1.2
jws==0.1.3
kombu==5.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.2.1
msgpack==1.0.5
mssql-django==1.3
multipledispatch==1.0.0
multidict==6.0.4
munkres==1.1.4
oauth2client==3.0.0
openai==1.3.2
# opencv-python==4.8.1.78
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.1
pandas==2.0.2
pandas-datareader==0.10.0
patsy==0.5.2
pep8==1.7.1
phonenumbers==8.13.11
Pillow==9.2.0
platformdirs==2.6.2
pluggy==1.2.0
ply==3.11
prompt-toolkit==3.0.39
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.7.0
# pytesseract==0.3.10 # computer vision related
pyodbc==5.0.1
pyparsing==3.1.0
PyQt5-sip==12.11.0
pytest==7.1.2
python-barcode==0.15.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2023.3
# PyYAML==6.0
queuelib==1.5.0
redis==4.5.5
requests-oauthlib==1.3.1
requests-ratelimiter==0.4.2
requests-toolbelt==0.7.0
rsa==4.9
sentry-sdk==1.15.0
sphinx==7.2.6
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tenacity==8.2.3
# tk==0.1.0
tomli==2.0.1
toolz==0.12.0
tzdata==2023.3
# ultralytics==8.0.207 # computer vision related
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
whitenoise==6.4.0
# wxPython==4.2.0
xhtml2pdf==0.2.11
yarl==1.9.2
yfinance==0.2.35
zict==2.1.0
zipp==3.16.1
sqlalchemy==2.0.25
finac==0.5.8
# zope.interface==6.0
